requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
attrdict==2.0.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
configparser==5.2.0
dataclasses==0.6
decouple==0.0.7
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.1
h11==0.13.0
idna==3.3
minio==6.0.2
psycopg2==2.9.3
pydantic==1.9.0
pyminio==0.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
